backend/requirements.txt
Django>=5.2,<6.1
django-cors-headers>=4.6,<5
django-environ>=0.12,<1
gunicorn>=23,<24
whitenoise>=6.8,<7
apify-client>=1.7,<2
google-generativeai>=0.8,<0.9
